move reference area into code-studio
